# moe quantization support int8 and fp8