@rasmith 2025-02-12
> > @vgokhale @Chi-Chu319 How long does it take to run these unit tests? Upstream vLLM usually asks for shorter unit tests because of CI. Would it be possible to get a set of parameterizations that can be ran in upstream vLLM but won't take a long time?
> 
> @rasmith I don't know. How long can we take in the CI? This is just the regular MoE kernel adapted to support int8. You can change use the same set of UTs as the current vllm MoE kernel.

30 seconds according to upstream